Add Google authentication in request.js
This commit introduces Google authentication in the request.js file. A new constant Google_URL has been added to store the authentication URL. A new function continueWithGoogle has been added to handle the Google authentication process. This function fetches the URL and handles any errors that may occur during the process. The function is then exported for use in other parts of the application.

Additionally, the Authorization token in the getRequest function has been commented out. This may need to be reviewed in future commits to ensure it doesn't affect other parts of the application.